🚴‍♀️ Move affiliations to own file
