+++ packages/frontmatter/src/index.ts
@@ -2,3 +2,4 @@ export * from './FrontmatterBlock';
 export * from './licenses';
 export * from './downloads';
 export * from './Authors';
+export * from './Affiliations';
